+++ ios/RNSentry.mm
@@ -47,10 +47,6 @@
 #import "RNSentryFramesTrackerListener.h"
 #endif
 
-@interface SentryTraceContext : NSObject
-- (nullable instancetype)initWithDict:(NSDictionary<NSString *, id> *)dictionary;
-@end
-
 @interface SentrySDK (RNSentry)
 
 + (void)captureEnvelope:(SentryEnvelope *)envelope;