+++ test/sdk.test.ts
@@ -432,9 +432,29 @@ describe('Tests the SDK functionality', () => {
       expect(actualIntegrations).toEqual(expect.not.arrayContaining([expect.objectContaining({ name: 'Spotlight' })]));
     });
 
-    it('adds spotlight integration', () => {
+    it('adds spotlight integration with enableSpotlight', () => {
       init({
         enableSpotlight: true,
+      });
+
+      const actualOptions = usedOptions();
+      const actualIntegrations = actualOptions?.integrations;
+      expect(actualIntegrations).toEqual(expect.arrayContaining([expect.objectContaining({ name: 'Spotlight' })]));
+    });
+
+    it('adds spotlight integration with spotlight bool', () => {
+      init({
+        spotlight: true,
+      });
+
+      const actualOptions = usedOptions();
+      const actualIntegrations = actualOptions?.integrations;
+      expect(actualIntegrations).toEqual(expect.arrayContaining([expect.objectContaining({ name: 'Spotlight' })]));
+    });
+
+    it('adds spotlight integration with direct url', () => {
+      init({
+        spotlight: 'http://localhost:8969/stream',
       });
 
       const actualOptions = usedOptions();